Fix more Linux build warnings.

+++ blakserv/saveall.c
@@ -45,7 +45,7 @@ INT64 SaveAll(void)
       We make our own copy since the time functions use a static
       buffer. */
    save_time = GetTime();
-   sprintf(time_str,"%lli",save_time);
+   sprintf(time_str,"%lli",(long long) save_time);
    
    save_ok = True;
 
@@ -96,15 +96,15 @@ void SaveControlFile(INT64 save_time)
    fprintf(savefile,"# Control file for last successful save\n");
    fprintf(savefile,"#\n");
    fprintf(savefile,"# Files written:\n");
-   fprintf(savefile,"# %s%s%lli\n",ConfigStr(PATH_LOADSAVE),GAME_FILE_SAVE,save_time);
-   fprintf(savefile,"# %s%s%lli\n",ConfigStr(PATH_LOADSAVE),ACCOUNT_FILE_SAVE,save_time);
-   fprintf(savefile,"# %s%s%lli\n",ConfigStr(PATH_LOADSAVE),STRING_FILE_SAVE,save_time);
-   fprintf(savefile,"# %s%s%lli\n",ConfigStr(PATH_LOADSAVE),DYNAMIC_RSC_FILE_SAVE,save_time);
+   fprintf(savefile,"# %s%s%lli\n",ConfigStr(PATH_LOADSAVE),GAME_FILE_SAVE,(long long) save_time);
+   fprintf(savefile,"# %s%s%lli\n",ConfigStr(PATH_LOADSAVE),ACCOUNT_FILE_SAVE,(long long) save_time);
+   fprintf(savefile,"# %s%s%lli\n",ConfigStr(PATH_LOADSAVE),STRING_FILE_SAVE,(long long) save_time);
+   fprintf(savefile,"# %s%s%lli\n",ConfigStr(PATH_LOADSAVE),DYNAMIC_RSC_FILE_SAVE,(long long) save_time);
    fprintf(savefile,"#\n");
    fprintf(savefile,"# Last successful save was at %s\n",TimeStr(save_time));
    fprintf(savefile,"#\n");
    fprintf(savefile,"\n");
-   fprintf(savefile,"LASTSAVE %lli\n",save_time);
+   fprintf(savefile,"LASTSAVE %lli\n",(long long) save_time);
    
    fclose(savefile);
 }
